requirements.txt
chromadb
langchain
langchain-community
sentence-transformers
beautifulsoup4
langchain_huggingface
langchain-openai
lxml
langchain_chroma
openai
fastapi
uvicorn
streamlit
PyMuPDF
PyJWT
passlib
pydantic[email]
python-multipart
